+++ backend/src/main/java/com/deliveranything/domain/settlement/service/dto/SettlementSummary.java
@@ -1,11 +1,9 @@
 package com.deliveranything.domain.settlement.service.dto;
 
-import java.math.BigDecimal;
-
 public record SettlementSummary(
-    BigDecimal totalTargetAmount,
-    BigDecimal totalPlatformFee,
-    long transactionCount
+    Long totalTargetAmount,
+    Long totalPlatformFee,
+    Integer transactionCount
 ) {
 
 }
